[WIP] - notifications and error fixing
