+++ app/src/main/java/com/ftn/eventhopper/activities/MainActivity.java
@@ -36,8 +36,10 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
         UserService.initialize(getApplicationContext());
+
+        setContentView(R.layout.activity_main);
+        //UserService.initialize(getApplicationContext());
         if (UserService.isTokenValid()) {
             ClientUtils.connectWebSocket();
         }
